refactor(editor): Remove old canvas code (no-changelog) (#13343)
