+++ packages/editor-ui/package.json
@@ -29,11 +29,6 @@
     "@codemirror/state": "^6.4.1",
     "@codemirror/view": "^6.26.3",
     "@fontsource/open-sans": "^4.5.0",
-    "@jsplumb/browser-ui": "^5.13.2",
-    "@jsplumb/common": "^5.13.2",
-    "@jsplumb/connector-bezier": "^5.13.2",
-    "@jsplumb/core": "^5.13.2",
-    "@jsplumb/util": "^5.13.2",
     "@lezer/common": "^1.0.4",
     "@n8n/api-types": "workspace:*",
     "@n8n/chat": "workspace:*",
