+++ packages/editor-ui/src/App.vue
@@ -100,22 +100,22 @@ watch(defaultLocale, (newLocale) => {
 				<BannerStack v-if="!isDemoMode" />
 			</div>
 			<div id="header" :class="$style.header">
-				<router-view name="header"></router-view>
+				<RouterView name="header" />
 			</div>
 			<div v-if="usersStore.currentUser" id="sidebar" :class="$style.sidebar">
-				<router-view name="sidebar"></router-view>
+				<RouterView name="sidebar" />
 			</div>
 			<div id="content" :class="$style.content">
 				<div :class="$style.contentWrapper">
-					<router-view v-slot="{ Component }">
-						<keep-alive v-if="$route.meta.keepWorkflowAlive" include="NodeViewSwitcher" :max="1">
+					<RouterView v-slot="{ Component }">
+						<KeepAlive v-if="$route.meta.keepWorkflowAlive" include="NodeView" :max="1">
 							<component :is="Component" />
-						</keep-alive>
+						</KeepAlive>
 						<component :is="Component" v-else />
-					</router-view>
+					</RouterView>
 				</div>
 				<div v-if="hasContentFooter" :class="$style.contentFooter">
-					<router-view name="footer" />
+					<RouterView name="footer" />
 				</div>
 			</div>
 			<div :id="APP_MODALS_ELEMENT_ID" :class="$style.modals">
