+++ packages/editor-ui/src/__tests__/defaults.ts
@@ -135,7 +135,6 @@ export const defaultSettings: FrontendSettings = {
 		enabled: false,
 		credits: 0,
 	},
-	betaFeatures: [],
 	easyAIWorkflowOnboarded: false,
 	partialExecution: {
 		version: 1,
