+++ packages/editor-ui/src/composables/useHistoryHelper.ts
@@ -7,7 +7,7 @@ import { useUIStore } from '@/stores/ui.store';
 
 import { onMounted, onUnmounted, nextTick } from 'vue';
 import { useDeviceSupport } from '@n8n/composables/useDeviceSupport';
-import { getNodeViewTab } from '@/utils/canvasUtils';
+import { getNodeViewTab } from '@/utils/nodeViewUtils';
 import type { RouteLocationNormalizedLoaded } from 'vue-router';
 import { useTelemetry } from './useTelemetry';
 import { useDebounce } from '@/composables/useDebounce';
