+++ packages/editor-ui/src/utils/typeGuards.ts
@@ -6,8 +6,6 @@ import type {
 } from 'n8n-workflow';
 import { nodeConnectionTypes } from 'n8n-workflow';
 import type { IExecutionResponse, ICredentialsResponse, NewCredentialsModal } from '@/Interface';
-import type { jsPlumbDOMElement } from '@jsplumb/browser-ui';
-import type { Connection } from '@jsplumb/core';
 import type { Connection as VueFlowConnection } from '@vue-flow/core';
 import type { RouteLocationRaw } from 'vue-router';
 import type { CanvasConnectionMode } from '@/types';
@@ -51,14 +49,6 @@ export const isCredentialModalState = (value: unknown): value is NewCredentialsM
 
 export const isResourceMapperValue = (value: unknown): value is string | number | boolean => {
 	return ['string', 'number', 'boolean'].includes(typeof value);
-};
-
-export const isJSPlumbEndpointElement = (element: Node): element is jsPlumbDOMElement => {
-	return 'jtk' in element && 'endpoint' in (element.jtk as object);
-};
-
-export const isJSPlumbConnection = (connection: unknown): connection is Connection => {
-	return connection !== null && typeof connection === 'object' && 'connector' in connection;
 };
 
 export function isDateObject(date: unknown): date is Date {
